Simple testing for the `predict` API.
